microservices/style-transfer-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aio-pika==9.3.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
cloudinary==1.36.0
requests==2.31.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
xformers>=0.0.22
opencv-python==4.8.1.78
Pillow==10.0.1
numpy>=1.24.0
datasets>=2.14.0
huggingface_hub>=0.17.0